WIP: gnutar: add `--allow-remote` flag, remove `--force-local`
This commit removes the `--force-local` flag and adds the
`--alow-remote` flag instead. For the most common use cases this should
be a sane default and most users shouldn't experience any differences.

This is still a WIP after a discussion on IRC about how to treat it. It
seems to be trivial enough to swap the behaviour (as implemented by this
commit). The missing pieces are probably ignoring --force-local if set
and potentially fixing the localization…

The conversation was brought up by `pie_` that linked to an debian bug
from 2005 [1].

Recently there was the very same question on the `help-tar` mailing list
[2] which didn't get much/any attention as to why this behaviour is
still the case.

An alternative to swapping the flag is removing/disabling the feature
entirely (and potentially provide a `gnutarRemote` package with the
standard behaviour). Another approach would be using an Open/FreeBSD
flavored tar but that would cause UX breakage...

The next steps will be deciding if anything at all should actually be
done and if so what...

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=290435
[2] https://lists.gnu.org/archive/html/help-tar/2018-03/msg00000.html