WIP: add dnssec-trigger-script
This will at some point be used to fix the openresolve/NM<>unbound
integration that is currently throwing errors whenever the DNS server is
changed.

I still have the suspicion that the actual problem is our usage of the
rc manager + dns=unbound in the NetworkManager configuration. Maybe we
have to tell openresolv to not use the dnssec-trigger-script? But it
might as well be that this package is really the only viable way right
now..