gnutls: remove autogen from build dependencies
There's an error when compiling autogen on macos Big Sur with #105026,
and it compiles fine without autogen, so I see no reason to keep it.

The dependency on autogen was originally introduced in 31a128b32bd12b5ebae,
but unfortunately there's no explanation for the reason and no linked issue.