nixos/systemd: introduce hardening profiles for services
The current state of the art in NixOS is to repeat some common hardening
options across each and every service individually. This leads to
inconsistencies across services that mostly depend on those that are
reviewing the changes.

With the NixOS module system we have the opportunity to hook into the
generation of all the systemd units and set some sane (hardening)
defaults.

This commit introduces a new
`systemd.services.<service-name>.defaultHarddefaultHardening` option
that allows specifiying a profile level that should be applied.

The meaning of a profile level is currently underspecified and could be
anything from a versioned subset of defaults to a list of presets for
variouds kinds of services (think webserver, background task, system
service, ...).

When treating them as versioned profiling levels we should stick to a
concept of *never* changing them (as long as a service is still
depending on them) and only ever add a new version that applies a
refined set of defaults. Refined could be anything from fixing issues
(too restrictive to be practical) to supporting new hardening features
that services should explicitly opt-in to by switching to a newer
profile.

Ideally, with this approach, we would end up with a lot of hardening
options being applied to almost all of our services. Thus eliminating a
bunch of potential security issues in our services.