cryptsetup: Specify a plugin directory that is available during runtime
On NixOS we don't have a global (stateful) directory for the crypsetup
plugins and thus we shouldn't try to create it during crypsetup install.
Instead we have a path that is available during runtime and points to
the cryptsetup plugins that systemd provides. At the time of this
writing that is a FIDO2 & TPM2 plugin.

This change also introduces the environment variable
NIXOS_CRYPTSETUP_EXTERNAL_TOKENS_PATH whcih can be used to override the
lookup directory during runtime. For example this is useful in our
initrd where we don't have the whole "toplevel" system tree available.