WIP: nixos/tests/installer: ensure that TPM works for disk encryption
This adds a test that verifies that our installations support using TPMs
for disk encryption (and decryption) from the installation onwards.

This test isn't a 100% representative of a real setup as it only depends
on the secure boot state of the system. In an actual deployment the
picked PCRs might depend on your personal threat model.